Fix Oxygen Distributor Gas Consumption Logic
Fixes issue where room size would not effect the rate of gas consumption
The old code had bugged logic resulting in it removing 1/mb every tick regardless of room size.

It seems like the old system was supposed to only run if 1000mb needed to be removed but it didn't work and ran every tick. I tried keeping it so it only removes 1000mb at a time but it feels bad from gameplay as the amount of gas in the distributor will drop in large chunks instead of constantly over time